requirements.txt
python-dotenv==1.0.1
litellm==1.17.3
llama-index
langchain
openai
gradio
transformers
unstructured[all-docs]==0.10.28
mpmath==1.3.0
tiktoken
requests
datetime
streamlit==1.31.0